Update README.md
In this project we have made a model which can convert gray scale images into colourised image using deep neural network and with the help of GAN(generative adversarial network). We used Cifar-10 Dataset to get the images needed for this project. It uses a CNN to analyze the colors across a set of color images, and their black and white versions.